Add backend and model switching controls
