+++ src/qwen_tui/tui/__init__.py
@@ -1,6 +1,7 @@
 """TUI components for Qwen-TUI."""
 
 from .app import QwenTUIApp, InputPanel, ThinkingWidget, ActionWidget
+from .backend_panel import BackendPanel
 from .chat_panel import ChatPanel
 
 __all__ = [
@@ -9,4 +10,5 @@ __all__ = [
     "ThinkingWidget",
     "ActionWidget",
     "ChatPanel",
+    "BackendPanel",
 ]
